Move SCopeServer class def. from scopeserver/__init__ to server_legacy.py
